rename boxes / Boxes --> box / Box consistently with trace type
