+++ src/traces/box/attributes.js
