+++ src/traces/box/layout_attributes.js
